type casting fixed
git-svn-id: https://svn.urbansim.org/src/trunk@19713 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ urbansim/models/scaling_agents_model.py
@@ -54,7 +54,7 @@ class ScalingAgentsModel(Model):
         
         for submodel in submodels:
             result[self.observations_mapping[submodel]] = self._simulate_submodel(submodel, 
-                                location_set, agent_set, agents_index)
+                                location_set, agent_set, agents_index).astype(result.dtype)
         return result
     
     def _simulate_submodel(self, submodel, location_set, agent_set, agents_index):
